[#6]pose_graph: reorganised project files and created pose_graph static library. Renamed pose_graph executable to pose_graph_node
