+++ pose_graph/include/pose_graph/keyframe.h
@@ -10,8 +10,7 @@
 #include "utility/tic_toc.h"
 #include "utility/utility.h"
 #include "parameters.h"
-#include "ThirdParty/DBoW/DBoW2.h"
-#include "ThirdParty/DVision/DVision.h"
+#include "FBrief.h"
 
 #define MIN_LOOP_NUM 25
