+++ pose_graph/include/pose_graph/utility/CameraPoseVisualization.h
