+++ pose_graph/src/keyframe.cpp
@@ -1,4 +1,7 @@
-#include "keyframe.h"
+#include "pose_graph/keyframe.h"
+
+#include "ThirdParty/DBoW/DBoW2.h"
+#include "ThirdParty/DVision/DVision.h"
 
 template <typename Derived>
 static void reduceVector(vector<Derived> &v, vector<uchar> status)
