+++ pose_graph/src/pose_graph_node.cpp
@@ -16,11 +16,13 @@
 #include <eigen3/Eigen/Dense>
 #include <opencv2/opencv.hpp>
 #include <opencv2/core/eigen.hpp>
-#include "keyframe.h"
-#include "utility/tic_toc.h"
-#include "pose_graph.h"
-#include "utility/CameraPoseVisualization.h"
-#include "parameters.h"
+
+#include "pose_graph/keyframe.h"
+#include "pose_graph/parameters.h"
+#include "pose_graph/pose_graph.h"
+#include "pose_graph/utility/CameraPoseVisualization.h"
+#include "pose_graph/utility/tic_toc.h"
+
 #define SKIP_FIRST_CNT 10
 using namespace std;
