+++ pose_graph/src/utility/CameraPoseVisualization.cpp
@@ -1,4 +1,4 @@
-#include "CameraPoseVisualization.h"
+#include "pose_graph/utility/CameraPoseVisualization.h"
 
 const Eigen::Vector3d CameraPoseVisualization::imlt = Eigen::Vector3d(-1.0, -0.5, 1.0);
 const Eigen::Vector3d CameraPoseVisualization::imrt = Eigen::Vector3d( 1.0, -0.5, 1.0);
